Don't punish every crate with serde-json
