+++ crates/test_utils/Cargo.toml
@@ -13,7 +13,6 @@ doctest = false
 # Avoid adding deps here, this crate is widely used in tests it should compile fast!
 dissimilar = "1.0.2"
 text-size = "1.0.0"
-serde_json = "1.0.48"
 rustc-hash = "1.1.0"
 
 stdx = { path = "../stdx", version = "0.0.0" }
